[unittest] specify softmax template type
Template type in activation functions needs to be specified to avoid
errors on ndk-build.

+++ Applications/ReinforcementLearning/DeepQ/jni/main.cpp
@@ -340,7 +340,7 @@ int main(int argc, char **argv) {
         nntrainer::sharedConstTensor test;
         try {
           in_tensor = nntrainer::Tensor(
-            {input}, {nntrainer::Tformat::NCHW, nntrainer::Tdatatype::FP16});
+            {input}, {nntrainer::Tformat::NCHW, nntrainer::Tdatatype::FP32});
         } catch (...) {
           std::cerr << "Error while construct tensor" << std::endl;
           return 0;
@@ -442,9 +442,9 @@ int main(int argc, char **argv) {
         nntrainer::Tensor q_in, nq_in;
         try {
           q_in = nntrainer::Tensor(
-            inbatch, {nntrainer::Tformat::NCHW, nntrainer::Tdatatype::FP16});
+            inbatch, {nntrainer::Tformat::NCHW, nntrainer::Tdatatype::FP32});
           nq_in = nntrainer::Tensor(next_inbatch, {nntrainer::Tformat::NCHW,
-                                                   nntrainer::Tdatatype::FP16});
+                                                   nntrainer::Tdatatype::FP32});
         } catch (...) {
           std::cerr << "Error during tensor constructino" << std::endl;
           return 0;
@@ -502,7 +502,7 @@ int main(int argc, char **argv) {
         nntrainer::Tensor in_tensor;
         try {
           in_tensor = nntrainer::Tensor(
-            inbatch, {nntrainer::Tformat::NCHW, nntrainer::Tdatatype::FP16});
+            inbatch, {nntrainer::Tformat::NCHW, nntrainer::Tdatatype::FP32});
           mainNet.forwarding({MAKE_SHARED_TENSOR(in_tensor)}, {Q});
           mainNet.backwarding(iter);
         } catch (...) {
